Implement publishers in a more consistent way
